+++ domain/Logic/Repositories/IUserRepository.cs
@@ -1,0 +1,9 @@
+using domain.Models;
+
+namespace domain.Logic.Repositories;
+
+public interface IUserRepository : IRepository<User>
+{
+    Result IsUserExists(string phoneNumber);
+    Result<User> GetUserByLogin(string phoneNumber);
+}
